+++ flag/service/namespaceauth/namespaceauth.go
@@ -1,6 +1,0 @@
-package namespaceauth
-
-type NamespaceAuth struct {
-	ViewAllTargetGroup     string
-	TenantAdminTargetGroup string
-}