+++ service/controller/rbac/resource/namespaceauth/delete.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 
 	"github.com/giantswarm/microerror"
+	pkgkey "github.com/giantswarm/rbac-operator/pkg/key"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
@@ -12,53 +13,60 @@ import (
 )
 
 func (r *Resource) EnsureDeleted(ctx context.Context, obj interface{}) error {
-	namespace, err := key.ToNamespace(obj)
+	ns, err := key.ToNamespace(obj)
 	if err != nil {
 		return microerror.Mask(err)
 	}
 
-	roles := []string{
-		"view-all",
-		"tenant-admin",
+	clusterRoles := []string{
+		pkgkey.OrganizationReadClusterRoleName(ns.Name),
 	}
 
-	for _, role := range roles {
+	for _, clusterRole := range clusterRoles {
 
-		_, err = r.k8sClient.RbacV1().RoleBindings(namespace.Name).Get(ctx, role, metav1.GetOptions{})
+		_, err = r.k8sClient.RbacV1().ClusterRoles().Get(ctx, clusterRole, metav1.GetOptions{})
 		if apierrors.IsNotFound(err) {
 			// do nothing
 		} else if err != nil {
 			return microerror.Mask(err)
 		} else {
-			r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("deleting %#q role binding", role))
+			r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("deleting %#q clusterrole", clusterRole))
 
-			err = r.k8sClient.RbacV1().RoleBindings(namespace.Name).Delete(ctx, role, metav1.DeleteOptions{})
+			err = r.k8sClient.RbacV1().ClusterRoles().Delete(ctx, clusterRole, metav1.DeleteOptions{})
 			if apierrors.IsNotFound(err) {
 				// do nothing
 			}
 			if err != nil {
 				return microerror.Mask(err)
 			} else {
-				r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("role binding %#q has been deleted", role))
+				r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("clusterrole %#q has been deleted", clusterRole))
 			}
 		}
+	}
 
-		_, err = r.k8sClient.RbacV1().Roles(namespace.Name).Get(ctx, role, metav1.GetOptions{})
+	roleBindings := []string{
+		pkgkey.WriteAllCustomerGroupRoleBindingName(),
+		pkgkey.WriteAllAutomationSARoleBindingName(),
+	}
+
+	for _, roleBinding := range roleBindings {
+
+		_, err = r.k8sClient.RbacV1().RoleBindings(ns.Name).Get(ctx, roleBinding, metav1.GetOptions{})
 		if apierrors.IsNotFound(err) {
 			// do nothing
 		} else if err != nil {
 			return microerror.Mask(err)
 		} else {
-			r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("deleting %#q role", role))
+			r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("deleting %#q role binding", roleBinding))
 
-			err = r.k8sClient.RbacV1().Roles(namespace.Name).Delete(ctx, role, metav1.DeleteOptions{})
+			err = r.k8sClient.RbacV1().RoleBindings(ns.Name).Delete(ctx, roleBinding, metav1.DeleteOptions{})
 			if apierrors.IsNotFound(err) {
 				// do nothing
 			}
 			if err != nil {
 				return microerror.Mask(err)
 			} else {
-				r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("role %#q has been deleted", role))
+				r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("role binding %#q has been deleted", roleBinding))
 			}
 		}
 	}