+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Move management of static resources from Helm into code.
+- Remove `view-all` related roles/bindings.
+- Bind customer admin group to `cluster-admin` cluster role in target organization namespace.
+
 ## [0.8.0] - 2020-11-19
 
 ### Added